requirement.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
jupyter>=1.0.0
ipykernel>=6.0.0
selenium==4.15.0
webdriver-manager==4.0.1 
streamlit >=1.50.0
webdriver-manager
